+++ aaem/__init__.py
@@ -1,3 +1,3 @@
-__version__ = '0.14.3+klukwan'
+__version__ = '0.14.5+klukwan'
 __url__ = 'https://github.com/gina-alaska/alaska_affordable_energy_model'
-__download_url__ = 'https://github.com/gina-alaska/alaska_affordable_energy_model/releases/tag/v0.14.3'
+__download_url__ = 'https://github.com/gina-alaska/alaska_affordable_energy_model/releases/tag/v0.14.5'