+++ aaem/forecast.py
@@ -86,6 +86,9 @@ class Forecast (object):
         self.yearly_total_kWh = DataFrame({"year":years,
                           "total":kWh['consumption'].values}).set_index("year")
         self.average_nr_kWh = kWh['consumption non-residential'].values[-3:].mean()
+        if np.isnan(self.average_nr_kWh):
+            temp = kWh['consumption non-residential']
+            self.average_nr_kWh = temp[ np.logical_not(np.isnan(temp))].mean() 
         self.yearly_nr_kWh = DataFrame({"year":years,
                           "total":kWh['consumption non-residential'].values}).set_index("year")
         #~ print self.average_nr_kWh
@@ -142,6 +145,7 @@ class Forecast (object):
         population = self.population.ix[idx].T.values[0]
         self.measured_consumption = self.yearly_total_kWh.ix[idx] 
         consumption = self.yearly_res_kWh.ix[idx].T.values[0]
+
         if len(population) < 10:
             self.diagnostics.add_warning("forecast", 
                   "the data range is < 10 matching years for "\
@@ -155,7 +159,7 @@ class Forecast (object):
             m, b = np.polyfit(population,consumption,1) 
         except TypeError:
             raise RuntimeError, "Known population & consumption do not overlap"
-        
+
         fc_consumption = (m * self.population + b) + self.average_nr_kWh
 
         start = int(self.measured_consumption.index[-1] + 1)
@@ -234,7 +238,8 @@ class Forecast (object):
             pass
     
         self.generation.columns = ["kWh generation"]
-            
+        
+        
     def forecast_generation_by_type (self):
         """
         forecasts the generation by each fuel type
@@ -601,7 +606,7 @@ class Forecast (object):
                            kWh_gen.round().astype(int), g_map] ,axis=1)
         except ValueError:
             self.electric_dataframe = None
-            
+
     
     
     def save_electric (self, csv_path, png_path):